requirements.txt
gradio==6.5.0
openai==2.15.0
chromadb==1.4.1
python-dotenv==1.2.1
tiktoken==0.12.0
